Updating route testcases for conda in originsTest file
