+++ test/lib/condaRepoAccess.js
@@ -4,9 +4,7 @@
 const { assert } = require('chai')
 const sinon = require('sinon')
 const proxyquire = require('proxyquire').noCallThru()
-const httpMocks = require('node-mocks-http')
 const requestPromiseStub = sinon.stub()
-const originCondaRoutes = require('../../routes/originConda')
 
 const fetchModuleStub = {
   callFetch: requestPromiseStub
@@ -19,16 +17,14 @@ const channelData = {
   packages: {
     'sample-package': { subdirs: ['linux-64'] },
     'another-package': {},
-    'sample-lib': {},
-    tensorflow: { subdirs: ['linux-64'] }
+    'sample-lib': {}
   },
   subdirs: ['linux-64']
 }
 
 const repoData = {
   packages: {
-    'pkg-1': { name: 'sample-package', version: '1.0', build: '0' },
-    'pkg-2': { name: 'tensorflow', version: '2.15.0', build: 'cuda120py39hb94c71b_3' }
+    'pkg-1': { name: 'sample-package', version: '1.0', build: '0' }
   }
 }
 
@@ -159,85 +155,5 @@ describe('CondaRepoAccess', () => {
       const matches = await condaRepoAccess.getPackages('conda-forge', 'sample')
       assert.deepEqual(matches, [{ id: 'sample-package' }, { id: 'sample-lib' }])
     })
-  })
-
-  describe('Conda Origin Routes', () => {
-    it('handles a valid GET request for revisions', async () => {
-      requestPromiseStub.onFirstCall().resolves(channelData)
-      requestPromiseStub.onSecondCall().resolves(repoData)
-      const request = createGetOriginCondaRevisionsRequest('tensorflow')
-      const response = httpMocks.createResponse()
-
-      const router = initializeRoutes()
-      await router._getOriginCondaRevisions(request, response)
-      assert.strictEqual(response.statusCode, 200)
-      assert.deepEqual(response._getData(), ['linux-64:2.15.0-cuda120py39hb94c71b_3'])
-      assert.isTrue(requestPromiseStub.calledTwice)
-    })
-
-    it('handles a valid GET request for package listings', async () => {
-      requestPromiseStub.onFirstCall().resolves(channelData)
-      const request = createGetOriginCondaRequest('conda-forge')
-      const response = httpMocks.createResponse()
-
-      const router = initializeRoutes()
-      await router._getOriginConda(request, response)
-      assert.strictEqual(response.statusCode, 200)
-      assert.deepEqual(response._getData(), [{ id: 'tensorflow' }])
-      assert.isTrue(requestPromiseStub.called)
-    })
-
-    it('returns a 404 error for a non-existent channel', async () => {
-      requestPromiseStub.onFirstCall().resolves(channelData)
-      const request = createGetOriginCondaRequest('tensor')
-      const response = httpMocks.createResponse()
-
-      const router = initializeRoutes()
-      await router._getOriginConda(request, response)
-      assert.strictEqual(response.statusCode, 404)
-      assert.strictEqual(response._getData(), 'Unrecognized Conda channel tensor')
-    })
-
-    it('returns a 404 error for a non-existent package in revisions', async () => {
-      requestPromiseStub.onFirstCall().resolves(channelData)
-      requestPromiseStub.onSecondCall().resolves(repoData)
-      const request = createGetOriginCondaRevisionsRequest('tensorflow1212')
-      const response = httpMocks.createResponse()
-
-      const router = initializeRoutes()
-      await router._getOriginCondaRevisions(request, response)
-      assert.strictEqual(response.statusCode, 404)
-      assert.strictEqual(response._getData(), 'Package tensorflow1212 not found in channel conda-forge')
-      assert.isTrue(requestPromiseStub.calledOnce)
-    })
-
-    function createGetOriginCondaRevisionsRequest(name) {
-      return httpMocks.createRequest({
-        method: 'GET',
-        url: `/origins/conda/'conda-forge'/linux-64/${name}/revisions`,
-        baseUrl: 'https://dev.clearlydefined.io',
-        params: {
-          channel: 'conda-forge',
-          subdir: 'linux-64',
-          name: name
-        }
-      })
-    }
-
-    function createGetOriginCondaRequest(channel) {
-      return httpMocks.createRequest({
-        method: 'GET',
-        url: `/origins/conda/${channel}/tensorflow/`,
-        baseUrl: 'https://dev.clearlydefined.io',
-        params: {
-          channel: channel,
-          name: 'tensorflow'
-        }
-      })
-    }
-
-    function initializeRoutes() {
-      return originCondaRoutes(condaRepoAccess, true)
-    }
   })
 })
