+++ test/routes/origins.js
@@ -6,7 +6,9 @@ const { expect, assert } = require('chai')
 const sinon = require('sinon')
 const proxyquire = require('proxyquire')
 const fs = require('fs')
+const httpMocks = require('node-mocks-http')
 const originMavenRoutes = require('../../routes/originMaven')
+const originCondaRoutes = require('../../routes/originConda')
 
 describe('Pypi origin routes', () => {
   let router
@@ -86,6 +88,125 @@ describe('Maven Origin routes', () => {
 
   function getResponse(filename) {
     return router._getSuggestions(loadFixture(`${fixturePath}/${filename}.json`))
+  }
+})
+
+describe('Conda Origin Routes', () => {
+  let condaRepoAccess
+  let cacheMock
+
+  const requestPromiseStub = sinon.stub()
+  const fetchModuleStub = {
+    callFetch: requestPromiseStub
+  }
+
+  const createCondaRepoAccess = proxyquire('../../lib/condaRepoAccess', {
+    './fetch': fetchModuleStub
+  })
+
+  const channelData = {
+    packages: {
+      tensorflow: { subdirs: ['linux-64'] }
+    },
+    subdirs: ['linux-64']
+  }
+
+  const repoData = {
+    packages: {
+      'pkg-2': { name: 'tensorflow', version: '2.15.0', build: 'cuda120py39hb94c71b_3' }
+    }
+  }
+
+  beforeEach(() => {
+    cacheMock = {
+      get: sinon.stub(),
+      set: sinon.stub()
+    }
+
+    condaRepoAccess = createCondaRepoAccess(cacheMock)
+  })
+
+  afterEach(() => {
+    sinon.restore()
+    requestPromiseStub.resetHistory()
+  })
+
+  it('handles a valid GET request for revisions', async () => {
+    requestPromiseStub.onFirstCall().resolves(channelData)
+    requestPromiseStub.onSecondCall().resolves(repoData)
+    const request = createGetOriginCondaRevisionsRequest('tensorflow')
+    const response = httpMocks.createResponse()
+
+    const router = initializeRoutes()
+    await router._getOriginCondaRevisions(request, response)
+    assert.strictEqual(response.statusCode, 200)
+    assert.deepEqual(response._getData(), ['linux-64:2.15.0-cuda120py39hb94c71b_3'])
+    assert.isTrue(requestPromiseStub.calledTwice)
+  })
+
+  it('handles a valid GET request for package listings', async () => {
+    requestPromiseStub.onFirstCall().resolves(channelData)
+    const request = createGetOriginCondaRequest('conda-forge')
+    const response = httpMocks.createResponse()
+
+    const router = initializeRoutes()
+    await router._getOriginConda(request, response)
+    assert.strictEqual(response.statusCode, 200)
+    assert.deepEqual(response._getData(), [{ id: 'tensorflow' }])
+    assert.isTrue(requestPromiseStub.called)
+  })
+
+  it('returns a 404 error for a non-existent channel', async () => {
+    requestPromiseStub.onFirstCall().resolves(channelData)
+    const request = createGetOriginCondaRequest('tensor')
+    const response = httpMocks.createResponse()
+
+    const router = initializeRoutes()
+    await router._getOriginConda(request, response)
+    assert.strictEqual(response.statusCode, 404)
+    assert.strictEqual(response._getData(), 'Unrecognized Conda channel tensor')
+  })
+
+  it('returns a 404 error for a non-existent package in revisions', async () => {
+    requestPromiseStub.onFirstCall().resolves(channelData)
+    requestPromiseStub.onSecondCall().resolves(repoData)
+    const request = createGetOriginCondaRevisionsRequest('tensorflow1212')
+    const response = httpMocks.createResponse()
+
+    const router = initializeRoutes()
+    await router._getOriginCondaRevisions(request, response)
+    assert.strictEqual(response.statusCode, 404)
+    assert.strictEqual(response._getData(), 'Package tensorflow1212 not found in channel conda-forge')
+    assert.isTrue(requestPromiseStub.calledOnce)
+  })
+
+  function createGetOriginCondaRevisionsRequest(name) {
+    return httpMocks.createRequest({
+      method: 'GET',
+      url: `/origins/conda/'conda-forge'/linux-64/${name}/revisions`,
+      baseUrl: 'https://dev.clearlydefined.io',
+      params: {
+        channel: 'conda-forge',
+        subdir: 'linux-64',
+        name: name
+      }
+    })
+  }
+
+  function createGetOriginCondaRequest(channel) {
+    return httpMocks.createRequest({
+      method: 'GET',
+      url: `/origins/conda/${channel}/tensorflow/`,
+      baseUrl: 'https://dev.clearlydefined.io',
+      params: {
+        channel: channel,
+        name: 'tensorflow'
+      }
+    })
+  }
+
+  function initializeRoutes() {
+    return originCondaRoutes(condaRepoAccess, true)
   }
 })
